api/metrics/requirements.txt
functions-framework>=3.5.0
google-cloud-firestore>=2.14.0
google-auth>=2.25.2
